requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
langgraph>=0.0.20
gradio>=4.19.0
typing-extensions>=4.7.0
